requirements-dev.txt
# --- Форматирование и линтинг ---
black>=23.3.0              # форматирование кода
isort>=5.12.0              # сортировка импортов
ruff>=0.0.380              # линтер и исправление ошибок

# --- Тестирование ---
pytest>=8.2.0
pytest-django>=4.5.2
pytest-cov>=4.1.0           # отчёты по покрытию тестами
factory-boy>=3.2.1          # генерация фиктивных данных для тестов

# --- Утилиты для разработки ---
ipython>=8.15.0             # удобная интерактивная консоль
django-extensions>=3.3.2    # полезные команды для разработки Django
debugpy>=1.6.7              # отладка с VS Code

# --- Optional: Работа с API и документацией ---
drf-spectacular>=0.28.0     # альтернатива drf-yasg для OpenAPI (можно использовать вместо drf-yasg)
